fix(training-data): derive s3Path server-side + fix Axiom name collision
Production evidence after PR #2203: 197 B2 training uploads in 6h, zero
register-file-location events, 6,635 download 404s. The registration
code path wasn't running for new uploads because users were testing on
stale client bundles (training pages live for hours during a run) that
don't yet send `s3Path` on the tRPC upsert — so the `backend === 'b2'
&& s3Path` guard failed silently.

Changes:
- Add `parseB2Url` helper in s3-utils. Primary check is the
  `*.backblazeb2.com` hostname pattern (env-free — works in admin
  scripts that run in pods without `S3_UPLOAD_B2_ENDPOINT` set).
  Falls back to matching the configured endpoint when it IS set, so
  custom proxies or non-public B2 fronts still parse.
- Controller: derive s3Path from the committed URL when the client
  didn't send one. Covers stale clients now and makes registration
  robust to client/server deploy skew going forward.
- Fix Axiom log name collision: `safeError()` returns `{name: 'Error'}`,
  so spreading it after the literal `name:` overwrote the event name.
  Every `resolve-download-url-failed` and `register-file-location-failed`
  event was landing under `name: "Error"` (6,635 in 6h). Spread
  `...safeError(err)` FIRST so the literal wins.
- Backfill: reuse the shared `parseB2Url` helper instead of a local copy.
